requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-google-genai
google-genai
langchain-text-splitters
faiss-cpu
google-generativeai